Move ErrorsView into its own folder
as a preparation for splitting the file up into files for each subcomponent
