+++ src/browser/modules/Stream/CypherFrame/ErrorsView/ErrorsView.test.tsx
