+++ src/browser/modules/Stream/SysInfoFrame/SysInfoFrame.tsx
@@ -29,7 +29,7 @@ import {
   StatusbarWrapper,
   StyledStatusBar
 } from '../AutoRefresh/styled'
-import { ErrorsView } from '../CypherFrame/ErrorsView'
+import { ErrorsView } from '../CypherFrame/ErrorsView/ErrorsView'
 import LegacySysInfoFrame from './LegacySysInfoFrame/LegacySysInfoFrame'
 import { SysInfoTable } from './SysInfoTable'
 import { InlineError } from './styled'
